modified:   src/server.c It seems that the message process and parent process are only allowed to write to sockets that they themselves have created. Therefore, creating a socket in a parent process and then trying to write to it in the child process results in a "bad file descriptor" error from write().
I show that this is the case by printing the file descriptors of the
sockets belonging to the parent and the child at the start. The fd of
the "receive messages" socket in the child process is the same as
the fd of one of the parent sockets, suggesting that the two processes
may only operate on sockets that they themselves have created.

This means that I'll have to go back to the address book format for now.